share common implementation
